% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce.R
\name{as.list.ndr::Dataset}
\alias{as.list.ndr::Dataset}
\title{Convert a Dataset to a list of data frames}
\usage{
\method{as.list}{`ndr::Dataset`}(x, ...)
}
\arguments{
\item{x}{A Dataset}

\item{...}{Ignored}
}
\value{
A named list of data.frames, one per data variable
}
\description{
Convert a Dataset to a list of data frames
}
