% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affine.R
\name{as_geotransform}
\alias{as_geotransform}
\title{Get the GDAL geotransform of an AffineIndex}
\usage{
as_geotransform(x)
}
\arguments{
\item{x}{An \link{AffineIndex}}
}
\value{
Named double vector of length 6, in GDAL order
}
\description{
Get the GDAL geotransform of an AffineIndex
}
