% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce.R
\name{as_variable}
\alias{as_variable}
\title{Quick Variable constructor from a named-dim array}
\usage{
as_variable(data, ...)
}
\arguments{
\item{data}{An R array}

\item{...}{Dimension names as \code{name = size} pairs (ignored, names used)
OR a character vector of dim names}
}
\value{
A Variable
}
\description{
Quick Variable constructor from a named-dim array
}
\examples{
# From a matrix with named dims
v <- as_variable(matrix(1:12, 3, 4), lat = 3, lon = 4)

}
