% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{$.ndr::Dataset}
\alias{$.ndr::Dataset}
\title{Extract a DataArray from a Dataset}
\usage{
\method{$}{`ndr::Dataset`}(ds, var_name)
}
\arguments{
\item{ds}{A Dataset}

\item{var_name}{Character name of the variable}
}
\value{
A DataArray
}
\description{
Use \code{ds$temperature} or \code{ds[["temperature"]]} to extract a variable
as a DataArray with its relevant coordinates attached.
}
