% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate.R
\name{coord_length}
\alias{coord_length}
\title{Get the length of a coordinate}
\usage{
coord_length(x)
}
\arguments{
\item{x}{A coordinate}
}
\value{
Integer length
}
\description{
Get the length of a coordinate
}
