% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate.R
\name{coord_values}
\alias{coord_values}
\title{Get the values of a coordinate}
\usage{
coord_values(x)
}
\arguments{
\item{x}{A coordinate (ImplicitCoord or ExplicitCoord)}
}
\value{
Vector of coordinate values
}
\description{
Get the values of a coordinate
}
