% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{dataarray_op}
\alias{dataarray_op}
\title{Apply an operation between Variables/DataArrays, preserving coordinates}
\usage{
dataarray_op(e1, e2, op)
}
\description{
Apply an operation between Variables/DataArrays, preserving coordinates
}
\keyword{internal}
