% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{ds_dims}
\alias{ds_dims}
\title{List dimension names and sizes across a Dataset}
\usage{
ds_dims(ds)
}
\arguments{
\item{ds}{A Dataset}
}
\value{
Named integer vector of all dimensions
}
\description{
List dimension names and sizes across a Dataset
}
