% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate.R
\name{ExplicitCoord}
\alias{ExplicitCoord}
\title{Create an explicit (irregular) coordinate}
\usage{
ExplicitCoord(dimension = character(0), values = NULL)
}
\arguments{
\item{dimension}{Dimension name (character, length 1)}

\item{values}{Vector of coordinate values}
}
\description{
Create an explicit (irregular) coordinate
}
\examples{
time <- ExplicitCoord(
  dimension = "time",
  values = as.Date("2020-01-01") + 0:364
)
coord_length(time)  # 365

}
