% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate.R
\name{ImplicitCoord}
\alias{ImplicitCoord}
\title{Create an implicit (regular-grid) coordinate}
\usage{
ImplicitCoord(
  dimension = character(0),
  n = integer(0),
  offset = numeric(0),
  step = numeric(0)
)
}
\arguments{
\item{dimension}{Dimension name (character, length 1)}

\item{n}{Number of elements (integer, length 1)}

\item{offset}{Start value (numeric, length 1)}

\item{step}{Step size (numeric, length 1)}
}
\description{
Create an implicit (regular-grid) coordinate
}
\examples{
lat <- ImplicitCoord(dimension = "lat", n = 180L, offset = -89.5, step = 1.0)
coord_values(lat)  # -89.5, -88.5, ..., 89.5
coord_lookup(lat, 0)  # integer index of latitude 0

}
