% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce.R
\name{names.ndr::Dataset}
\alias{names.ndr::Dataset}
\title{Variable names in a Dataset}
\usage{
\method{names}{`ndr::Dataset`}(x)
}
\arguments{
\item{x}{A Dataset}
}
\value{
Character vector of variable names
}
\description{
Returns all available variable names, including lazy (not-yet-loaded)
variables from the backend.
}
