% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy.R
\name{collect}
\alias{collect}
\title{Materialise a lazy object by reading data from disk}
\usage{
collect(x, ...)
}
\arguments{
\item{x}{A LazyDataArray (or DataArray, which is returned as-is).}
}
\value{
A DataArray with data in memory.
}
\description{
Materialise a lazy object by reading data from disk
}
