% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate.R
\name{coord_dim}
\alias{coord_dim}
\title{Get the dimension name of a coordinate}
\usage{
coord_dim(x, ...)
}
\arguments{
\item{x}{A coordinate}
}
\value{
Character scalar
}
\description{
Get the dimension name of a coordinate
}
