% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable.R
\name{ndim}
\alias{ndim}
\title{Get the shape of a Variable}
\usage{
ndim(x, ...)
}
\arguments{
\item{x}{A Variable}
}
\value{
Named integer vector of dimension lengths
}
\description{
Get the shape of a Variable
}
